listing.dom and listing.cdom are now integers. I believe listing.dom is valid for inactive listings. But for active ones we still dont have a way to figure it out.
This commit implements this logic on API side so if we have to change this logic again (quite likely), we dont have to make client change again.
So here's the logic:
If the listing is inactive, return the DOM stored in db.
If its active, then (now - list_date) shows up.
fixes #513