These past 2 commits made me feel like a damn genius! They fix the problem that Facebook shares were not branded.
Brand.getCurrent() was supposed to use current user's brand.
However, it was not working, because it relied on process.domain.brand to work

process.domain.brand was being set on a middleware defined by Brand's controller.
That middleware was using req.user, which was supposed to be defined by auth helper (app.auth.bearer)

But since app.auth.bearer was not a middleware, but a wrapper function, it meant that Brand controller's middleware was being executed first.

That means when Brand middleware was being called, req.user was still not there.

The patch is not like this:

All endpoints now go through optional passport authentication.
So calling app.auth.optionalBearer is not necessary anymore. Its going to be deprecated.
Instead, all user authentication is actually happening on a middleware.
And this middleware is executed earlier than brand's middleware.

I had to think about the proper fix to this solution for like 3 days. It seems obvious to you know, but trust me, it wasnt.