When the domain is rollbacked or commited, take the db connection from it. Ticks that keep running after a rollback could've used the connection outside the transaction which was dangerous.
Basically something like this was happening:

BEGIN
C1;
C2;
C3; <-- This one was throwing an error
ROLLBACK;
C4; <- This was was queued in before the rollback and there's no way to stop it.
       And since it's outside of any transaction, it's gonna be commited anyways.

This makes sure that when a domain is done it wont have access to the db connection anymore.