[Contact] Fix index attribute when migrating from old contacts data
Previously, when address fields were broken down to their components,
there was no way for clients to know which components belong together.

This change adds index field while applying the migration script,
and uses a trick to easily compute the index for all attributes per
contact and attribute_type.

It also adds a unique constraint on contact, attribute_def and index
so that no two attributes of the same type get same index on the same
contact, in case clients send bogus inputs to the endpoints.