This commit makes some fundumental changes to the notification model. It somehow paves the way for future changes to simplify the model and get rid of the stupid object subject whatever model.
For now, what it does is the following:

1. The notification message is something we generated differently per user and per transport.
Also, it only makes sense to do so when the notification is made, as the object/subjects
are not immutable. Therefore, the generated message is valid only for a specific time.

Therefore, we now store the following when the notification is created (and not when it's gonna be sent)

* notification_user.sms_message
* notification_user.push_message
* notification_user.message

2. notifications.data has been added as a JSONB field.
I personally hate JSONB fields. But in this case it makes sense.

The idea is that instead of storing object/subjects, the notification generators can just
send the data that the templates need. The templates will be rendered based on provided data.

The data, jsonb object, will only act as a log, from then on.

Also it's free-form.

3. notification_user.branch is also added. It stores the branch link we create for sms links.
Since we now generate the text messages when the notification is created, we also create the branch link then.

It's not optimal. But meh.

This is all done to pave the way for product#328