Nastiest bug ever.
Calling `done` several times causes the "The connection has been closed by other party" error.
It also messed up with the the next person who is acquiring the connection. Maybe the second time we're calling it, it already belongs
to someone else, and we close it for him!

This fixes several issues. Hopefully, we wont leak database connections anymore.