[MLS] Lock listing table at the beginning of Listing.create
This will prevent the race condition between two concurrent workers
trying to update two revisions of the same listing as observed in a
couple of instances where the older revision would replace the newer
one.

We hope that this change will prevent at least a class of outdated
listing issue causes.