During investigation of web#2137 and web#2138 I realized some context values were missing from deal_context but they were set of current_deal_context.
More investigation shows that alpine's deal_context is empty and most of
boer's deal_context is removed too.

I suspect a migration may have deleted them which must be investigated further.

For now Im apploying this contraint to production to make sure whatever triggered
such tyranny would not be able to do so on production.