Prepare to make a change we always wanted to do: Dropping listings_filters Materialized View. Refreshing it is a huge hassle and puts too much load on the db.
Basically a series of optimizations and cleanups can be done, listed as follows:

Dropping unused users.address_id to make sure addresses is solely used to properties and listings.
Adding a trigger to addresses to updated listings_filters so we can get rid of the materialized view.
Adding a trigger to create mls_context table instead of depending on it as a function
Creating a table that unions current_deal_context and mls_context instead of calculating it on the fly every single time
Making calendar a table instead of a view