Right now we're expericing timeouts when ios app for users like me or try is opened up. That's because our users have access to root brands which have thousands of deals.
On Web it's not a problem because only backoffce users have access to these brands
and those users are supposed to use the inbox endpoint which only returns submitted items.

But iOS has no notion of BackOffice vs. Agent. It only is curated for agents.
Therefore, it asks for all the deals under your brand, which for me and Trey would be thousands.

Those endpoint calls make the server use a ton of memory and evenrually case timeout errors.

So this patch basically limits the number of deals a client can request to 300.