[Contact/List] A major refactor on lists
The main motivation behind this change was to fix tag filters of a list
when a tag is renamed. This way users won't have to fix thier lists
manually.

I also experimented with a different way for
bulk INSERTing multiple using `json_populate_recordset` for `filters`.
The update script also takes care of `filters` itself, so there's no
need for a separate model or script.

1. Removed json fields for `filters` and `args`.
   `filters` is now a separate table, and `args` is just a field on the
   main table.
2. Both tables for lists and list members have been renamed.
3. Create endpoint now responds with the complete list object.
4. The method for creating default stage lists is fixed.
5. `is_pinned` field is dropped from lists. It wasn't being used.
6. A new `is_editable` field is added with proper write access control
   to support future system lists if needed.