This lays the foundation for giving the ability to admins to create deals based on deleted agents.
The problem was, so far, we were actually deleting the user-brand relation.
But now we're using the deleted_at pattern to make sure that the connection
is not lost.

This means we'll have the ability to offer special UI to admins that
still shows such agents and allows them to be picked.

I will still have to implement endpoints that allows admins to do so
but since we dont have a product decision on the UI yet, any API
I create may end up being premature.

Also that'd be a tiny amount of work. Therefore Im pospoting that
to the time we have a direction for product.

for product#532