So, here's a long one.
The problem we're facing here is that each email can receive many events.
An email campaign sent to 2K people will have thousands of events.

We store this events in their almost raw form in emails_events table.
Clients, in various places, need to show information about those emails.
For example in Marketing Insights. But also in various other places like CRM.

Obviously, exposing clients to all events is not feasible.
Specially in a page like Marketing Insight's index page where it shows tens of
email campaigns in an index. We'd have to expose tens of thousands of rows of emails_events
which is obviously out of question.

Therefore, we need to keep some numbers.
Now, due to several reasons, the numbers have to be kept on 3 different tables:

email_campaigns
emails
email_campaing_emails

So we'd have emails.opened, which shows how many people have opened this email.
The initial implementation of this was simple an elegant (Just like the first Matrix).
It would increase the number of occurances of each event every time something happens.

So, each time an "opened" event happed, we'd basically do
emails.opened++
email_campaigns.opened++
email_campaign_emails.opened++

The problem, however, was that on emails and email_campaigns, we needed the
counts to be unique per recipient.

In another words, that simple elegant logic answered the question
"how many times this email has been opened"
But the question we wanted to answe was
"how many recipients opened this email"

That means we cannot just +1 events. We'd have to see if they have happened before.

This means calculation. I hate calculation. That's error prone.

So, the second solution was this:

Every time an event happened, a trigger was executed, which recounted all events
for all emails in that campaign.

So basically on each new event, all records were counted again.
So it basically had to sequentially scan thousands of rows and then _update_ thousands
of rows.

Yes. That's very expensive. And yes, I was a naive fool.

The moment the first user sent a campaign of thousands of emails, we started to receive
hunreds of email events. And each one of them triggered a whole rebuild.

Which meant all rows went into a deadlock and endpoints started to timeout.
We could not even record stats.

So, this is the third solution. This is the way it works:

Every time a new event is added, we'll just insert it's record into emails_events table.
Then, we'll queue the email campaign to be updated.

Every time there's a 30 second inactivity (no events), we'll THEN update the campaign.

I know, the stupid campaign rebuild thing _is_ expensive and must be avoided.
But tehre's several reasons I'm going this:

(btw if you read it until here and you're judging me, fuck you.)

Also, we're thinking of adding a capability to system to send notifications
to users, letting them know about results of their campaign.
"35 more people opened your email, 3 clicked on it, 2 bounced"

The method which rebuilds a campaign is the perfect place to put that notification.

Now, there _is_ a cheap way to update the rows without rebuilding
the whole campaign and updating thousands of rows.
And you're asking why I did that instead of this hideous react-inspired method.

Of course I do realize that it's not a good idea to UPDATE all those rows all the time.
It'll put pressure, etc, etc.

The reason I did this is that at this stage I'm concerned about the system being correct rather
that fast. We don't have too many users, campaign and emails.
The important thing is being correct. And this is the the easiest way to get it correct.
Also, every time we realize something is wrong with the rebuild function (update_email_campaign_stats)
I can just fix it and run it on all campaigns, Voila.

Now, if you are a critical and level headed reader you'd ask:
What fucking "calculation" are you talking about that you say is complicated and error-prone
that you're running away from with all these stupid tricks.

Or in another words, why the simple elegant approach #1 could not be updated
so instead of a dumb +1, we could do a +1 if it was unique.

Here's the problem.

Lets say I have 1 single query in my database that inserts the emails_events record,
then goes sees previous records in the database for that recipient, and finally +1
the relevant records only if it was unique.

In Postgres, in order to do all those steps in 1 action you'd have to use CTE's.
But here's the problem:

When CTE's write, other CTE's don't have information to the updated table.
It's as if each CTE is being executed on a different transaction.

That means I cannot have 1 query which i a single atomic action updates everything.

I'd have to use Javascript/PLPSQL to insert, then read, then update.

1. Insert Record
2. Get current number
3. Update it

That's very prone to all kind of race conditions and shit.
Also, I'd still need a "rebuild" thing anyways, because if that procedural thing
has any issues, we should be able to fix the logic and recount things.

So basically, the method would be error-prone due to race conditions
and we'd need to have update_email_campaing_stats anyways in order to maintain and support this.

That means 2 different executions of the same functionality. And they are different. And getting them
both perfect would be difficult.

So I'd rather have 1 expensive source of truth for the moment and rely on that. Until later, when
(hopefully) scaling challenges being, and we'll do more.

Of course there's also the approach of storing the data in both
Postgres and other cheaper stores like Redis and having an eventually-consistend approach. That'd be
what the Twitters and Google's of the world would do. But that will be really error pront and not something
I'm willing to maintain with a team of this size.