Implemented new deal role fields that belong to office.
The most important thing here is the introduction of deal_role.office_address.

We've been talking about a more robust way of dealing with addresses in Rechat.
This is the begining of that initiative.

We are not relying on Postgres stdaddr composite data type to store an address.
A helper function named STADDR_TO_JSON is provided to make it easy to input
an address from JS to Postgres.

Next steps would be to provide an endpoint that uses standardize_address
to parse an address and offer it to clients as a helper method.

Right now we use ROW_TO_JSON to convert an staddr from Postgrea to JS.
However, we may be able to extend that and provide another helper function
which would assemble `street_address` and `full_address` of the address as well
so clients can just use that.

If this implementation is successful, I will migrate
deal_role.current_address and deal_role.future_address to this format as well.

My only concern was that the address columns, and therefore our responses,
we going to be heavily deflated as the staddr has to many columns.

But using the amazing JSON_STRIP_NULLS we remove them and the responses are as small
as they could be.