We LEFT JOIN deals_roles with brokerwolf_agents. The problem is, brokerwolf_agents may end up returning multiple entries for this, if the agent has 2 matching MLSID's on the same account. This means for a given deal role id, the query will return 2 rows.
This is obviously incorrect behavior which is fixed by the patch.
However, it hasn't showed up as an issue anywhere yet.

This was discovered during investigating server#1312 from Ish from BrokerWolf.
This issue meant that we were sending 2 agents to BW instead of one.

However, and unfortunately, fixing this did not resolve the BW issue.
Although it had to be fixed anyways.