Right now, when we are ready to switch to a new context, we do a `context.enter`.
The problem with this approach is, the main event loop is not supposed to have any context.
So when we do a `context.enter` to go to the new context, it's the main event loop that
switches context.

So basically this happens:

Main Event loop Has no context
	Request1 comes in
	Main Event Loop has no context itself. It does a context.enter() on context C1
	Context becomes c1 on the main event loop
	Middleware on Request1 runs through C1

	Request2 comes in
	Main Event Loop is at C1 itself. It does a context.enter() on context C2
	It seems that C2 is branched-off of C1 because Main Event Loop had switched to C1 before

We'll have to move the error management and database handling code to Context because it's copied a lot.
But this patch fixes the problem for the moment.

Hopefully this also resolves the 500 Request Aborted issues we've had as well.