So, we used to use matrix_unique_id as Foreign Key for mls-driven relationship. Now that we're going multi-mls, we reazied that mui is unique only per-each-mls.
Therefore, we decided to get rid of it's usage and have proper
uuid based foreign keys. Problem with that, however, is that there's
depdendencies between those relationships.

I'm going to use photos <> listings relationship as an example
but this applies to other cases as well.

We had 3 choices:
1. To set photos.listing during photo insertion.
That means if the listing wasn't inserted before, this would be left
empty and when the listing came up, the connection would'nt be made.

2. To set NOT NULL on photos.listing. Then, the photo insertion would fail
until we finally have the listing.
Rrelying on worker's failures is something I'm not going to bet on at this stage.

3. To create a trigger so every time we insert a listing, we find all it's listings
and update photos.listings for them all.
This is an interesting case but I'm not really ready for it yet as of now. We are doing
too many new things at the same time these days.

4. To go back to joining dynamically on (mls, mui).
Performance wise this is the slowest as it does joins during read operations
but it's th emost reliable for the moment.

I suspect in the long run we're going to do 3. But for the moment, I need reliability not performance.