As you can see we already have a join on deals_role as dr But here we make another join on deals_roles as drr
The reason is, we want to say:
Filter out those deals which he has no access to (using dr)

But among those, also filter out the ones with a matching role.

If we reuse dr for this part of query, then it'd apply both conditions
to all deals.

But we want the conditions to be applied separately.

Imagine a deal having 2 roles:
BuyerAgent A John Smith
Buyer      B Alice Mannings

If the query says: Filter out deals belonging to John Smith
that match the string `Alice`, then the query would fail
if we reuse dr condition for both as we don't have a role that
both belongs to John Smith and also matches `Alice`.

So we need to join them separately so we can filter them separately.

fixes server#1340