Here goes a long one which fixes #1343.
As explained in 1343, Docusign has a feature that allows you to sign
an envelope by printing it, signing all documents and pages and
uploading/faxing a dead pdf back to docusign.

Lets say we had an envelope with 2 documents:
  1. Document A, Originated from Form A, belonging to Task A.
  2. Document B, Originated from File B, belonging to Task B.

This envelope is sent to 3 users.
  1. User Alpha who digitally signs it.
  2. User Beta  who signs it on paper.
  3. User Theta who digitally signs it.

When 2 happens, Docusign generates a pdf including all pages from both documents A and B.
Users Beta signs them and uploads them back in a single dead pdf.

At this point, our envelope signatures are broken into 3 documents:
  Document A, signed by users Alpha and Theta
  Document B, signed by users Alpha and Theta
  Signed Copy, which is a dead PDF of documents A and B and signed by user Beta

So, docusign automatically adds the third document to our envelope and marks
the recipient as Signed.

So now Rechat is showing '3 of 3 Signed', on Task A and Task B.
When backoffice clicks on documents A and B, they see Alpha and Theta's
signatures but not Beta's. As Beta's signature was in 3rd document which
never existed in Rechat.

So, Rechat will now realize that a new document is added to the envelope
and whats to start tracking it. The question is,
which (task, submission) or (task, file) does it belong to?

The answer is, all of them. As the signed copy includes all pages of all documents.

But the envelope_document -> task is a 1:N relationship. Therefore I cannot
add an envelope_document that belongs to all tasks.
Making it a N:N relationship is too big of a task and complication which may
undermine the reliability we have here, which I'm not willing to risk.

Therefore, the current solution is, when this happens, I will add an activity to all tasks
and add the Signed copy pdf as a file to the room with a small activity log.

Unfortunately the docusign API will not give us much information of who added the file so it's
kind of vague and the whole solution is not perfect. But for the moment it stops the leak.

Since this is a very very rare occurance, the solution is enough for now.