Up until now, a user had access to all templates he had access to at the same time. So if we had 6 templates, 2 being avail to all, 2 being avail to brand A and 2 avail for brand B, a user who had access to both brands saw all 6 of them at the same time.
Also, now that we are actually gonna use template's
brand-limitations more seriously, as we have more clients, we need to start QA'ing these,
and the current behavior makes it really difficult to QA, as most testers have access to all brands.

So with these changes, in order to see templates of a brand, you have to activate that brand.

I renamed the endpoint to `GET /brands/:brand/templates`.
But we already have been using that endpoint for email templates.

So I renamed all those endpoints.