for #1355.
So far we had the following recipient types:
1. Tag
2. List
3. Individual Email (+ Contact)

Now I had to implement 2 new ones:
1. To send to all contacts
2. To send to all agents of a brand

Until now we determined the types based on column values but that was becoming too much hassle.
So I needed to add a new excplicit column.
I wanted to name it `recipient_type` but we were already using that for `TO, CC, BCC`.

So I
1. Renamed `recipient_type` to `send_type`
2. Added the new `recipient_type`
3. Added bunch of constraints to prevent malformed input
4. Wrote migrations
5. Wrote queries to convert new types to emails (`email/campagign/emails.sql`)