[Contact] Implement default tags logic
This includes a migration to add default tags to brands lacking them,
and also making sure brands are created with those default tags.

Another change has been made in the rename logic, to silently merge
tags if user is renaming a tag to an existing one. Previously this
operation resulted in a conflict exception.
The final change is regarding uniqueness check of tags, which used to
be case-sensitive, but now it's case-insensitive so users can't create
`Tag` and `tag` for a single brand; only one will be possible.