[Calendar] Fix an ugly performance issue with a hack!
The problem began when we introduced 4 new sources of calendar events,
two for each of Outlook and Google threads. These new subqueries use the
DISTINCT ON clause to fetch only the first email of a
thread. And things begin to get ugly right here.

DISTINCT ON clauses are a deal breaker when WHERE conditions are being
pushed down to subqueries because it may change the results. As a result
none of our conditions were being pushed down even to other parts of
the calendar view! So all of a sudden the entire calendar data would
load and then conditions where applied to it.

Now, there is an exception to to the rule about DISTINCT ON clauses:
The conditions (or quals) that are included in the DISTINCT ON clause
are safe to be pushed down. This is still a big limitation, specially
that most of the subquery's select fields in our case are constant! One
would've expected Postgres to sort it out somehow.

Nonetheless, the fact that most fields are constant in our case and
others have a hierarchical relationship (like brand), helped us here to
just add those fields the DISTINCT ON clause and be done with it.
This way, conditions will still get pushed to subqueries.