Our logic for who has access to a template was this so far: 1. If you have specifically gained access to it (on brands_allowed_templates table) 2. The template is global (eg noone has specific access to it on brands_allowed_templates)
Now, BFSIR, has been asking for some of the global templates to be removed from their brand.

With this change, we simplifi it:
You only gain access to a template if you have specifically been added to bat table.

Also, writing a migration that gives access to all current global templates to all brokerages.