Ability to render nunjucks templates in a pool of worker processes.
When we send a big campaign (several thousand templates), we're going
render thousands of templates. Since rendering is cpu bound, it
blocks the event loop. That means everything will be blocked until
rendering is done.

This process is happening in worker because EmailCampaign.sendDue is there.
This means everything in a worker script will be completely blocked, which will
cause a load of issues.

Also, the processing happens super slowly as it's not distributed between
multiple processes. By having a worker pool, we'll distribute it.