A complete refactor. The previous versions used to get an array of all contexts and render them all together. That mean there was an array that would have to contain all render requests. This array would grow large and large until the Nodejs' memory heap size was reached and nodejs would crash.
This new mechanism is far more scale-able. It creates a resource
handler which can be used to generate as many renders as possible
but they don't need to happen in 1 function call:
each render request is handled independantly so the GC can remove them.

Also did some cleanups. Also we don't need a pool anymore.
Each render request will spawn a process and then let it go when render is
completed.