fixes server#1462
So, the problem is, we are sending emails like this:

{
...,
headers: {}
}

But the mailgun-js API doesn't accept headers like that
it only accepts headers in this format:

{
...,
h:Header-Name: Header-Value
}

And if we pass any object to it, it will break with that error.
(https://github.com/mailgun/mailgun-js/issues/7)

The reason it was working until now is:
We always had headers:null
And therefore we never sent the headers: {} to it
Even where we want to send headers right now (like messages/email.js)
we are sending headers using h:Reply-To: <value> format and we don't use the jsonb headers field.

That's also changed now I made sure we now use the persisted method.