Clean up, dehardcore. Polish.
And here's a complete documentation on the SAML process and this implementation:

=== Providers ===

In SAML SSO, 2 (important) parties are involved.

1. Identity Provider (In our current case, Realogy)
2. Service Provider  (Rechat)

So the user logs into an IdP, then with a single-click, get's logged in into SP.

So, Rechat, is an SP in this case.

The way SAML works, it requires some configuration to be shared between SP and IdP (Like a signing certificate).

There are 2 ways we could implement the SAML:

1. In a way that it totally depends on 1 single IdP (Helped by Passport-Saml.Strategy)
2. In a way that an endpoint can support any IdP    (Helped by Passport-Saml.MultiStrategy)

My implementation uses Passport-Saml.Multistrategy to implement an IdP Agnostic SAML SP.
The advantage of method 2 is, later, if we decide to allow more service providers (Like Google and Microsoft)
we won't have to do any development.

That also means we'll have to store the configuration of each IdP we plan to support.
That information is stored in `sso_providers`.

=== Users ===

Imagine this scenario:

* We have emil@rechat.com as admin
* Realogy decides they want to hack into Rechat and gain admin access
* Realogy creates a user on their own system with email emil@rechat.com
* The hacker logs into Realogy with user emil@rechat.com and clicks on Rechat
* Rechat logs the hacker in as emil@rechat.com with full access

This is an actual possibility in a SAML Integration which trusts the IdP blindly.
But we cannot allow this to happen. Maybe for IdP's like Google and Microsoft, but not for Realogy.

To prevent that from happening, we cannot rely on email address provided
by IdP's as the _only_ login information.
But they do provide additional `userid` and `issuer` per SAML standard.

So basically, a user coming from an IdP, needs to be isolated to that IdP.

So these 3 users:

1. emil@rechat.com issued by Realogy IdP
2. emil@rechat.com issued by Google
3. emil@rechat.com a normal user on Rechat

Should not mix up during SSO. Otherwise, that's a security flaw.

In order to allow this to happen, we have a new table called `sso_users`
which basically stored the `userid, issuer` (coming from IdP) and the rechat-user-id.

When a user does a SAML login, we won't even look at their email address. We find
the corresponding rechat-user based on incoming `(userid,issuer) pair`

If the user doesn't exist, we will create a new shadow user with the profile (first name, last name, email) provided
by the IdP.

There's 1 flaw though:

In Rechat, `users.email` is unique.

Therefore, if there is already a user with the email address you're doing SSO with,
and that user doesn't belong to this IdP, the operation with fail.

In the future, we will be able to detect the situation, offer the user to login with his
existing user account, and _connect_ the SSO with the existing account.
(Only if he proves he owns that account by entering password)

However, for the initial implementation, we're skipping this part, as it'll require E2E development.

=== Sending users to Webapp ===

After the SSO is completed, we need to send the user to Webapp.

* If this is the first time the user is doing an SSO, we need to finalize his account (Claim shadow user)
* Otherwise, we need to log user in

This process looks like this:

1. If this is a new user, we'll reuse the UserActivation branch link.
2. If this is an existing user, we'll create him a new refresh token.
This refresh token, which is valid only for 5 minutes, will be used by client to
retrieve a new pair of `(access,refresh)` tokens.

An important point, however, is this:

Server _needs_ to generate a token here. But, each generated token in Rechat, requires to be
connected to a `client`. Up until now, our clients were Webapp and iOS App.

The way I resolved this issue is this:
Each IdP can contain the client-id which will be used for tokens on that IdP.
But they can reuse the same client.

The idea is, for now, we'll have a new client named `SAML Authenticator`. All `SAML Tokens` will be generated for that client.