for server#1155
HerokuFix used to rely on sending a new `\n` every once in a while to prevent heroku-router from closing the connection after 30s.
The problem was, prepending a `\n` on an xlsx file which is a binary file, corrupts it.
So we couldn't enable it.

Now we're relying on HTTP Headers to keep the Heroku Router's connection open.
Therefore, it can work for xlsx files as well.
So I'm enabling it for the report.