Using Orm.populate() to populate all would cause a maximum stack error in node. We also don't need checklists at all.
The only side effect is we dont have brands anymore. So we cannot extract office name.
We can still Orm.populate() the whole brand tree but I don't thik that's gonna be necessary.
The feature was broken for many months. Noone asked for it. At least we now have most of it back.
Except for office name.

fixes #1155