There are 2 fixes here. 1. When a deal is created with no address, the tasks created for that deal will have names like this: `: Information About Brokerage Services`
That's because we want the format `Address: Task` but the address is empty in those cases.
That's easily resolved by relying on deal.title now, which does actully have short address (if available).

The second problem is, in many cases, we do update deal titles. But then the room titles get outdated.
This resolves that issue using a beautiful trigger.

These have been there for years, but Sarah from League emailed me about them a couple of days ago.