workers: slow down on pushing peanar jobs
after some investigation, I found that the event loop gets locked up due
to the aggressive manner by which we push peanar jobs to rabbitmq.

One naive but effective approach is to just remove Promise.all and
do the job in a for loop and await on enqueuing calls.
my experiments show that if the payload is not more than a few KBs,
this approach is quite fast (~30s for 100,000 messages of 3 KBs,
compared to ~6s with the aggressive approach), while also maintaining
stability of the event loop, and garbage collection jobs.

when payload is larger than a few KBs, (or maybe larger than a few amqp
frames), both the framing work that happens in amqplib, as well as
the garbage collection operations required take their toll on the node
process, sometimes pushing it into an unrecoverable state in extreme cases.