trigger: fix cloning campaign
We were using Orm.populate to get email campaign and its recipients,
but contact is an association of the recipient model so it wasn't being
populated. This change will remove the usage of Orm.populate and simply
gets the recipients of the campaign directly from its model. This way
the contact ids won't come as null.