Big one, for #2148.
So, basically, we want to add task-level access controls. So task.acl is added. This is easy.
But then, these acl are based on _roles_. This is where it gets complicated. Current roles are
`BackOffice` and `Agents`. That means every time we are returning a set of tasks, we need to know
if the user accessing them is a BackOffice user or an Agent. But this is not as easy as you may
think. There's levels and levels of access here.

Also, the `BackOffice` and `Agents` are only for now. We are thinking of extending this to client portals,
offer portals, etc.

So, how do we handlde this complexity in different parts of the app?

The current solution creates a table-returning function: deals_acl(user)
This would return a table with following columns: deal, brand, acl

Basically, given a user id, it would return a list of all deals the user has access to with their:

* deal id
* brand id in which they have access to that deal through
* acl type

This function is then used in 4 places:

`Deal.limitAccess()`, which is used through js to make sure a user wont access a deal he doesn't have access to

Deal.filter() which instead of having a separate logic for deal acl's reused the function and logic (it used to copy the code)

Deal.getBrandDeals through deal/brand.sql which loads a list of deals for a brand

Task.getAll() through task/get.sql, which ensure when we return a set of tasks, only the ones the user has access to
are returned to him.

Now, up until now, first 3 items had separate implementations of the same logic. They copied code.
Now, thanks to the new table returning function, all 3 of them share the logic and code. Great.

The bad news? It has a bit of a performance impact.
The function, although is inlined, uses `brand_children()`, which is somewhat expensive. Not sure how I can eliminate that.

This new function, deals_acl, calls user_brands() 4 times. user_brands() calls brand_children(). So brand_children() ends up being called 4 times.
The previous code was calling brand_children() twice already. So the impact is not going to be huge.
Also note that the function takes only $25ms on our biggest brand with ~9000 sub brands so far. So the impact is really minimal.

Last note. This is sensitive stuff. Probably requires some serious testing.