Alright. Initially the idea of the ORM was this: Model Deal Checklist tells you I have an association to Model Tasks 1,2,3. Then you go fetch Tasks 1,2,3 and if there's any 1 of them missing, you throw an error. This was initially helpful during development of the Orm as it helped us catch a few issues. However, for a long long time I haven't seen Orm being useful for this.
Now this is what happened: On the Task model we are introducing an extra layer of ACL.
So now it's looking like this:
Deal Checklist tells you I have an association to Tasks 1,2,3.
Then you fo fetch Tasks 1,2,3 and you may get 1,3, but not 2, because you don't have access to 2.

Now, in this case, old code would throw an error. But with this commit, it would consider that
a graecful situation and just skip over B2.

Now, the question is, if the user doesn't have access to Task 2, shouldn't be filtered out
by Deal Checklist already? Why should Deal Checklist tell you to go get Tasks 1,2,3 when user doesn't
have access to 2?

That is a _very_ good question, and, honestly, im personally still not sure if what I'm doing here
is the best approach. But here's my thinking:

We can have 3 approaches to this:
1. Deal Checklist should do the ACL control for tasks. Because it's the one actually deciding
what tasks myst be returned to user.

Problem with it is security: If there's _any_ other association/surface to access tasks, they may
not implement this, and tasks may leak to users.

2. Tasks should do the ACL control for themselves. This is what this commit allows to happen.

3. Both should. If both do, then we wouldn't have to amend Orm and there's no concerns, except:

A. Database costs for these ACL checks is not cheap. So there's a performance impact.
B. I don't want the burden of development/maintenance to be twice as high (once for deal checklist,
and once for tasks themselves)

Now, everything I said about Deal Checklists here is purely and example. We may have more cases like this.
I want the Orm to allow this pattern for Deal Checklist -> Tasks relationship as well as any upcoming
future cases.